move updateOrgUnitWithLastUpdatedData to projectDb.ts
